backend/requirements/development.txt
-r base.txt

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
pytest-asyncio==0.25.3
factory-boy==3.3.1
faker==33.3.1
pytest-xdist==3.5.0

# Linting & Formatting
ruff==0.9.6

# Type Checking
mypy==1.14.1
django-stubs==5.1.3
djangorestframework-stubs==3.15.3

# Debugging
django-debug-toolbar==4.4.6
django-extensions==3.2.3
ipython==8.31.0

# Security Scanning
bandit==1.8.3
pip-audit==2.7.3
